Move RunLumiEventAnalyzer to FWCore/TestModules
It is used outside of FWCore, and it feels generic-enough to be in
TestModules.
