+++ FWCore/TestModules/plugins/BuildFile.xml
@@ -1,6 +1,8 @@
+<use name="DataFormats/Common"/>
 <use name="DataFormats/Provenance"/>
 <use name="FWCore/AbstractServices"/>
 <use name="FWCore/Framework"/>
+<use name="FWCore/MessageLogger"/>
 <use name="FWCore/ParameterSet"/>
 <use name="FWCore/PluginManager"/>
 <use name="FWCore/Sources"/>
