+++ FWIO/RNTupleTempTests/test/testLooperEventNavigation_cfg.py
@@ -19,7 +19,7 @@ process.source = RNTupleTempSource(
     ]
 )
 
-from FWCore.Framework.modules import RunLumiEventAnalyzer
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test = RunLumiEventAnalyzer(
     verbose = True
     , expectedRunLumiEvents = [
