+++ FWIO/RNTupleTempTests/test/testRunMergeCOPY1_cfg.py
@@ -28,7 +28,7 @@ process.source = RNTupleTempSource(
     , noEventSort = False
 )
 
-from FWCore.Framework.modules import RunLumiEventAnalyzer
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test = RunLumiEventAnalyzer(
     verbose = True,
     expectedRunLumiEvents = [
