+++ FWIO/RNTupleTempTests/test/testRunMergeMERGE3_cfg.py
@@ -41,7 +41,7 @@ process.thingWithMergeProducer = ThingWithMergeProducer()
 from FWIO.RNTupleTempOutput.modules import RNTupleTempOutputModule
 process.out = RNTupleTempOutputModule(fileName = 'testRunMergeMERGE3.root')
 
-from FWCore.Framework.modules import RunLumiEventAnalyzer
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test = RunLumiEventAnalyzer(
     verbose = True,
     expectedRunLumiEvents = [
