+++ FWIO/RNTupleTempTests/test/testRunMergeTEST1_cfg.py
@@ -38,7 +38,7 @@ process.source = RNTupleTempSource(
 from FWIO.RNTupleTempOutput.modules import RNTupleTempOutputModule
 process.out = RNTupleTempOutputModule(fileName = 'testRunMergeRecombined1.root')
 
-from FWCore.Framework.modules import TestMergeResults, RunLumiEventAnalyzer
+from FWCore.Framework.modules import TestMergeResults
 process.test = TestMergeResults(
 
     #   Check to see that the value we read matches what we know
@@ -154,6 +154,7 @@ process.test = TestMergeResults(
    ]
 )
 
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test2 = RunLumiEventAnalyzer(
     verbose = True,
     expectedRunLumiEvents = [
