+++ IOPool/Input/test/test_reduced_ProcessHistory_end_cfg.py
@@ -20,7 +20,7 @@ process.maxEvents.input = -1
 from IOPool.Output.modules import PoolOutputModule
 process.output = PoolOutputModule(fileName = 'reduced_test.root')
 
-from FWCore.Framework.modules import TestMergeResults, RunLumiEventAnalyzer
+from FWCore.Framework.modules import TestMergeResults
 process.testmerge = TestMergeResults(
     #   Check to see that the value we read matches what we know
     #   was written. Expected values listed below come in sets of three
@@ -133,6 +133,7 @@ process.source = PoolSource(
     duplicateCheckMode = "noDuplicateCheck"
 )
 
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test = RunLumiEventAnalyzer(
     verbose = True,
     expectedRunLumiEvents = [
