+++ IOPool/Tests/test/testRunMergeMERGE1_cfg.py
@@ -42,7 +42,7 @@ process.thingWithMergeProducer = ThingWithMergeProducer()
 from IOPool.Output.modules import PoolOutputModule
 process.out = PoolOutputModule(fileName = 'testRunMerge_a.root')
 
-from FWCore.Framework.modules import RunLumiEventAnalyzer
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test = RunLumiEventAnalyzer(
     verbose = True,
     expectedRunLumiEvents = [
