+++ IOPool/Tests/test/testRunMergePickEventsx_cfg.py
@@ -26,7 +26,7 @@ process.source = PoolSource(
     , eventsToProcess = ['1:1:2-1:1:7']
 )
 
-from FWCore.Framework.modules import RunLumiEventAnalyzer
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test = RunLumiEventAnalyzer(
     verbose = True,
     expectedRunLumiEvents = [
