+++ IOPool/Tests/test/testRunMergeTEST6_cfg.py
@@ -19,7 +19,7 @@ process.source = PoolSource(
     duplicateCheckMode = 'noDuplicateCheck'
 )
 
-from FWCore.Framework.modules import TestMergeResults, RunLumiEventAnalyzer
+from FWCore.Framework.modules import TestMergeResults
 process.test = TestMergeResults(
     #   Check to see that the value we read matches what we know
     #   was written. Expected values listed below come in sets of three
@@ -74,6 +74,7 @@ process.test = TestMergeResults(
     ]
 )
 
+from FWCore.TestModules.modules import RunLumiEventAnalyzer
 process.test2 = RunLumiEventAnalyzer(
     verbose = True,
     expectedRunLumiEvents = [
